examples/example_11_webhook/python_telegram_bot/requirements.txt
python-telegram-bot==21.7
